renamed unmatching files and added new dynamic route
